Reduce cross-platform build warning noise (#315)
* Reduce CI/build warning noise and harden cross-platform behavior.

Rebase warning-cleanup work onto current main, keep vdf_client session parser refactor from #313, and apply targeted warning/robustness fixes across CI workflows, CMake/build scripts, and core/runtime code paths.


* Restore full 2weso CI durations and tighten Python/Cargo target plumbing.

Remove temporary shortened 2weso test overrides in vdf-client-hw CI and align build scripts with modern CMake/Cargo target variables to eliminate unused PYTHON_EXECUTABLE warnings and keep MSVC-only CRT overrides scoped correctly.


* Consolidate checked integer casts across C/C++ and Python bindings.

Replace ad-hoc narrowing checks and release-elided asserts with a shared checked_cast helper to keep overflow handling consistent across wrapper and test paths.


* Fix unsigned-to-signed checked_cast overflow guard and add regression coverage.

The unsigned-to-signed conversion path now checks against the signed destination max, preventing silent wraparound for values above INT32_MAX, and adds a GoogleTest regression to lock this behavior.


* Fix signed-widening behavior in checked_cast and add regression coverage.

Avoid truncated bounds checks for same-signed widening casts and add positive/negative int32-to-int64 tests.


* Use checked_cast for primetest and verifier discriminant conversions.

Replace manual narrowing checks/casts with checked_cast to keep overflow handling consistent and centralized.


* Use 64-bit recursion and proof lengths in verifier APIs.

Align the Python and C bindings with the native verifier interface by accepting full-size proof blob lengths and recursion depth instead of narrowing to int32, removing wrapper-level truncation guards.


* Use logical OR for uint128 comparison operators.

Switch boolean composition in <= and >= from bitwise OR to logical OR for clearer intent and short-circuit behavior.


* Update verifier proof-length tests and known-proof fixture.

Switch verifier and stress tests to pass proof blob lengths as size_t without narrowing casts, and refresh verifier_test's fixed proof vector to a current-format fixture so the known-proof assertion remains meaningful with the widened verifier API.


* Mirror production verifier path in known-proof test.

Update verifier_test to construct and reuse the discriminant before calling CheckProofOfTimeNWesolowski directly, so the test reflects production verifier usage while keeping the current-format known-proof fixture.


* Pass witness_type directly to uint64 depth API.

Remove the obsolete int32_t intermediate cast in stress_test so the call matches the updated verifier depth parameter type.


* Inline env_flag use in prover_test fast-mode check.

Remove the redundant env_truthy wrapper and call env_flag directly to keep the test setup path simpler without changing behavior.


* Revert "Use logical OR for uint128 comparison operators."

This reverts commit 8db997ddb92685e3446821d52a881587a02b3f64.

---------

+++ src/c_bindings/c_wrapper.cpp
@@ -1,4 +1,5 @@
 #include "c_wrapper.h"
+#include "../checked_cast.h"
 #include <vector>
 #include <gmpxx.h>
 #include "../verifier.h"
@@ -9,7 +10,7 @@ extern "C" {
     bool create_discriminant_wrapper(const uint8_t* seed, size_t seed_size, size_t size_bits, uint8_t* result) {
         try {
             std::vector<uint8_t> seed_vector(seed, seed + seed_size);
-            integer discriminant = CreateDiscriminant(seed_vector, size_bits);
+            integer discriminant = CreateDiscriminant(seed_vector, checked_cast<int>(size_bits));
             mpz_export(result, NULL, 1, 1, 0, 0, discriminant.impl);
             return true;
         } catch (...) {
@@ -20,7 +21,7 @@ extern "C" {
     ByteArray prove_wrapper(const uint8_t* challenge_hash, size_t challenge_size, const uint8_t* x_s, size_t x_s_size, size_t discriminant_size_bits, uint64_t num_iterations) {
         try {
             std::vector<uint8_t> challenge_hash_bytes(challenge_hash, challenge_hash + challenge_size);
-            integer discriminant = CreateDiscriminant(challenge_hash_bytes, discriminant_size_bits);
+            integer discriminant = CreateDiscriminant(challenge_hash_bytes, checked_cast<int>(discriminant_size_bits));
             form x = DeserializeForm(discriminant, x_s, x_s_size);
             std::vector<uint8_t> result = ProveSlow(discriminant, x, num_iterations, "");
 
@@ -38,7 +39,7 @@ extern "C" {
         try {
             integer discriminant;
             mpz_import(discriminant.impl, discriminant_size, 1, 1, 0, 0, discriminant_bytes);
-            
+
             return CheckProofOfTimeNWesolowski(
                 -discriminant,
                 x_s,
